refactor(lab-3088): rename some methods and folders
`list_models` -> `models`
`get_model` -> `model`
`update_model` -> `update_properties_in_model`
`list_project_models` -> `project_models`
`get_project_models` -> `project_model`

`adapters/kili_api_gateway/model_configuration` -> `adapters/kili_api_gateway/llm`
