+++ src/kili/adapters/kili_api_gateway/kili_api_gateway.py
@@ -6,7 +6,7 @@ from kili.adapters.kili_api_gateway.asset.operations_mixin import AssetOperation
 from kili.adapters.kili_api_gateway.cloud_storage import CloudStorageOperationMixin
 from kili.adapters.kili_api_gateway.issue import IssueOperationMixin
 from kili.adapters.kili_api_gateway.label.operations_mixin import LabelOperationMixin
-from kili.adapters.kili_api_gateway.model_configuration.operations_mixin import (
+from kili.adapters.kili_api_gateway.llm.operations_mixin import (
     ModelConfigurationOperationMixin,
 )
 from kili.adapters.kili_api_gateway.notification.operations_mixin import (
